Extract util functions
